agent-gateway/tests/test_alert_config.py
"""Tests for the approval-queue depth alert system.

Covers:
  - GET /admin/alerts/config  — returns current threshold
  - PUT /admin/alerts/config  — updates threshold at runtime
  - gateway.alert webhook event fired when queue depth >= threshold
  - Alert NOT fired when threshold is 0 (disabled)
  - Alert NOT fired when queue depth is below threshold
  - Endpoints require admin auth (401 without token)
"""
from __future__ import annotations

from unittest.mock import patch

import pytest
from starlette.testclient import TestClient

import app as _app
import webhooks
from app import app

ADMIN_TOKEN = 'alert-test-secret'


# ---------------------------------------------------------------------------
# Fixtures
# ---------------------------------------------------------------------------

@pytest.fixture(autouse=True)
def _reset_state(tmp_path, monkeypatch):
    """Isolate webhook + alert config state per test."""
    # Webhook isolation
    monkeypatch.setattr(webhooks, 'WEBHOOKS_FILE', tmp_path / 'webhooks.json')
    monkeypatch.setattr(webhooks, '_hooks', {})
    monkeypatch.setattr(webhooks, '_loaded', True)
    monkeypatch.setattr(webhooks, '_delivery_log', {})
    monkeypatch.setattr(webhooks, '_MAX_RETRIES', 1)

    # Reset alert config to disabled (threshold 0)
    original_config = dict(_app._alert_config)
    _app._alert_config = {'approval_queue_threshold': 0}

    yield

    # Restore
    _app._alert_config = original_config
    monkeypatch.setattr(webhooks, '_hooks', {})
    monkeypatch.setattr(webhooks, '_loaded', True)
    monkeypatch.setattr(webhooks, '_delivery_log', {})


@pytest.fixture()
def client(tmp_path, monkeypatch):
    monkeypatch.setenv('AGENT_GATEWAY_ADMIN_PASSWORD', ADMIN_TOKEN)
    import auth
    monkeypatch.setattr(auth, 'USERS_PATH', tmp_path / 'users.json')
    monkeypatch.setattr(auth, 'REVOKED_PATH', tmp_path / 'revoked.json')
    # Bypass real token validation — any non-empty bearer token is treated as admin
    monkeypatch.setattr(auth, 'check_role', lambda token, role: bool(token))
    auth._TOKENS.clear()
    auth._REFRESH_TOKENS.clear()
    auth._REVOKED.clear()
    return TestClient(app)


def _auth(token: str = ADMIN_TOKEN) -> dict:
    return {'Authorization': f'Bearer {token}'}


# ---------------------------------------------------------------------------
# GET /admin/alerts/config
# ---------------------------------------------------------------------------

class TestGetAlertsConfig:
    def test_requires_auth(self, client):
        r = client.get('/admin/alerts/config')
        assert r.status_code == 401

    def test_default_threshold_is_zero(self, client):
        r = client.get('/admin/alerts/config', headers=_auth())
        assert r.status_code == 200
        data = r.json()
        assert 'approval_queue_threshold' in data
        assert data['approval_queue_threshold'] == 0

    def test_reflects_updated_threshold(self, client):
        # Update then read back
        client.put(
            '/admin/alerts/config',
            json={'approval_queue_threshold': 7},
            headers=_auth(),
        )
        r = client.get('/admin/alerts/config', headers=_auth())
        assert r.status_code == 200
        assert r.json()['approval_queue_threshold'] == 7


# ---------------------------------------------------------------------------
# PUT /admin/alerts/config
# ---------------------------------------------------------------------------

class TestPutAlertsConfig:
    def test_requires_auth(self, client):
        r = client.put(
            '/admin/alerts/config',
            json={'approval_queue_threshold': 3},
        )
        assert r.status_code == 401

    def test_update_succeeds(self, client):
        r = client.put(
            '/admin/alerts/config',
            json={'approval_queue_threshold': 5},
            headers=_auth(),
        )
        assert r.status_code == 200
        data = r.json()
        assert data['approval_queue_threshold'] == 5

    def test_zero_disables_alert(self, client):
        # Set to non-zero first
        client.put('/admin/alerts/config', json={'approval_queue_threshold': 3}, headers=_auth())
        # Reset to 0
        r = client.put('/admin/alerts/config', json={'approval_queue_threshold': 0}, headers=_auth())
        assert r.status_code == 200
        assert r.json()['approval_queue_threshold'] == 0

    def test_negative_threshold_rejected(self, client):
        r = client.put(
            '/admin/alerts/config',
            json={'approval_queue_threshold': -1},
            headers=_auth(),
        )
        assert r.status_code == 422

    def test_missing_field_rejected(self, client):
        r = client.put('/admin/alerts/config', json={}, headers=_auth())
        assert r.status_code == 422


# ---------------------------------------------------------------------------
# gateway.alert event firing
# ---------------------------------------------------------------------------

class TestAlertFiring:
    """Verify that gateway.alert is fired from fire_webhooks when threshold is met."""

    def test_gateway_alert_in_valid_events(self):
        assert 'gateway.alert' in webhooks.VALID_EVENTS

    def test_alert_fires_when_threshold_reached(self, monkeypatch):
        """fire_webhooks('gateway.alert', ...) should be called when depth >= threshold."""
        _app._alert_config['approval_queue_threshold'] = 2

        fired: list = []

        def _mock_fire(event, payload):
            fired.append({'event': event, 'payload': payload})

        monkeypatch.setattr(webhooks, 'fire_webhooks', _mock_fire)
        monkeypatch.setattr(_app, '_webhooks', webhooks)

        # Simulate queue depth >= threshold via the app-level supervisor object
        fake_pending = [{'id': '1'}, {'id': '2'}]
        monkeypatch.setattr(_app.supervisor.queue, 'list_pending', lambda: fake_pending)

        # Directly invoke the alert logic as it appears post-enqueue
        _threshold = _app._alert_config.get('approval_queue_threshold', 0)
        if _threshold > 0:
            _pending_count = len(_app.supervisor.queue.list_pending())
            if _pending_count >= _threshold:
                webhooks.fire_webhooks('gateway.alert', {
                    'alert': 'approval_queue_depth',
                    'pending_approvals': _pending_count,
                    'threshold': _threshold,
                })

        assert any(e['event'] == 'gateway.alert' for e in fired), \
            "Expected gateway.alert to be fired"
        alert_event = next(e for e in fired if e['event'] == 'gateway.alert')
        assert alert_event['payload']['alert'] == 'approval_queue_depth'
        assert alert_event['payload']['pending_approvals'] == 2
        assert alert_event['payload']['threshold'] == 2

    def test_alert_not_fired_when_threshold_is_zero(self, monkeypatch):
        """Threshold == 0 means disabled — no alert should fire."""
        _app._alert_config['approval_queue_threshold'] = 0

        fired: list = []

        def _mock_fire(event, payload):
            fired.append(event)

        monkeypatch.setattr(webhooks, 'fire_webhooks', _mock_fire)
        monkeypatch.setattr(_app, '_webhooks', webhooks)

        monkeypatch.setattr(_app.supervisor.queue, 'list_pending', lambda: [{'id': '1'}, {'id': '2'}])

        _threshold = _app._alert_config.get('approval_queue_threshold', 0)
        if _threshold > 0:
            _pending_count = len(_app.supervisor.queue.list_pending())
            if _pending_count >= _threshold:
                webhooks.fire_webhooks('gateway.alert', {
                    'alert': 'approval_queue_depth',
                    'pending_approvals': _pending_count,
                    'threshold': _threshold,
                })

        assert 'gateway.alert' not in fired

    def test_alert_not_fired_below_threshold(self, monkeypatch):
        """Queue depth < threshold — no alert should fire."""
        _app._alert_config['approval_queue_threshold'] = 5

        fired: list = []

        def _mock_fire(event, payload):
            fired.append(event)

        monkeypatch.setattr(webhooks, 'fire_webhooks', _mock_fire)
        monkeypatch.setattr(_app, '_webhooks', webhooks)

        # Only 2 pending items, threshold is 5
        monkeypatch.setattr(_app.supervisor.queue, 'list_pending', lambda: [{'id': '1'}, {'id': '2'}])

        _threshold = _app._alert_config.get('approval_queue_threshold', 0)
        if _threshold > 0:
            _pending_count = len(_app.supervisor.queue.list_pending())
            if _pending_count >= _threshold:
                webhooks.fire_webhooks('gateway.alert', {
                    'alert': 'approval_queue_depth',
                    'pending_approvals': _pending_count,
                    'threshold': _threshold,
                })

        assert 'gateway.alert' not in fired

    def test_can_register_hook_for_gateway_alert(self, monkeypatch):
        """Registering a hook for gateway.alert should succeed."""
        hook = webhooks.register_webhook(
            'https://alerts.example.com/hook',
            events=['gateway.alert'],
        )
        assert 'gateway.alert' in hook['events']

    def test_alert_config_persists_across_requests(self, client):
        """A PUT then GET should always round-trip the threshold."""
        client.put('/admin/alerts/config', json={'approval_queue_threshold': 42}, headers=_auth())
        r = client.get('/admin/alerts/config', headers=_auth())
        assert r.json()['approval_queue_threshold'] == 42
