agent-gateway/requirements.txt
fastapi>=0.95,<1
uvicorn[standard]>=0.22
jsonschema>=4.21.0,<5
httpx>=0.24
pytest>=7.4
beautifulsoup4>=4.12
lxml>=4.9
keyring>=23.8
requests>=2.31
pyyaml>=6.0
